Crate fabric_contract_macros

Definitions of the macros used for the Contracts.

There are macros for use with the Contract Structs, and the Transaction Functions

Attribute Macros

contract_impl
Use this macro to mark the implementation of the your contract structure

property
Define the properties of the datatype

transaction
Use this to mark the functions that are considered to be transaction functions

Derive Macros

DataTypeMacro
Use this to mark the structs that serve as complex data types Need to provide example